Enable API versioning
The backend API needs versioning. This is not to be done by Django to
avoid needless complexity and dependence upon Django but by the reverse
proxy / API Gateway instead.

The existing backend code is now confined to the v1 directory. New
versions can be added in the future and do not require the use of Django
or any specific version of it.